fix: clean up providers
0. Pulls message conversion utils into format module
1. Overhauls the provider tests which were flakey in CI
2. Removes moderation references for now, as we are investigating how to
   bring them in without the false positives
3. Removes cost tracking, as we don't want to keep up to date with the
   pricing details. We will track tokens instead
