--- crates/goose/.env.example
@@ -1,9 +1,0 @@
-# OpenAI Configuration
-OPENAI_API_KEY=your_api_key_here
-
-# Databricks Configuration
-DATABRICKS_HOST=your_databricks_host
-DATABRICKS_TOKEN=your_databricks_token
-
-# Test Configuration
-INTEGRATION_TESTS=true  # Set to true to run integration tests
