+++ crates/goose/src/agents/capabilities.rs
@@ -1,6 +1,5 @@
 use chrono::{DateTime, TimeZone, Utc};
 use mcp_client::McpService;
-use rust_decimal_macros::dec;
 use std::collections::HashMap;
 use std::sync::Arc;
 use std::sync::LazyLock;
@@ -193,11 +192,6 @@ impl Capabilities {
                     e.usage.total_tokens = Some(
                         e.usage.total_tokens.unwrap_or(0) + usage.usage.total_tokens.unwrap_or(0),
                     );
-                    if e.cost.is_none() || usage.cost.is_none() {
-                        e.cost = None; // Pricing is not available for all models
-                    } else {
-                        e.cost = Some(e.cost.unwrap_or(dec!(0)) + usage.cost.unwrap_or(dec!(0)));
-                    }
                 })
                 .or_insert_with(|| usage.clone());
         });
