+++ crates/goose/src/providers/formats/mod.rs
@@ -1,0 +1,3 @@
+pub mod anthropic;
+pub mod google;
+pub mod openai;
