+++ crates/goose/src/providers/mod.rs
@@ -4,19 +4,13 @@ pub mod configs;
 pub mod databricks;
 pub mod errors;
 pub mod factory;
-pub mod mock;
-pub mod model_pricing;
+pub mod formats;
+pub mod google;
+pub mod groq;
 pub mod oauth;
 pub mod ollama;
 pub mod openai;
-pub mod openai_utils;
+pub mod openrouter;
 pub mod utils;
 
-pub mod google;
-pub mod groq;
-pub mod openrouter;
-
 pub use factory::get_provider;
-
-#[cfg(test)]
-pub mod mock_server;
