+++ crates/goose/src/providers/openrouter.rs
@@ -4,15 +4,12 @@ use reqwest::Client;
 use serde_json::Value;
 use std::time::Duration;
 
-use super::base::{Moderation, ModerationResult, Provider, ProviderUsage, Usage};
+use super::base::{Provider, ProviderUsage, Usage};
 use super::configs::ModelConfig;
-use super::model_pricing::cost;
-use super::model_pricing::model_pricing_for;
 use super::utils::{emit_debug_trace, get_model, handle_response};
 use crate::message::Message;
-use crate::providers::openai_utils::{
-    check_openai_context_length_error, create_openai_request_payload_with_concat_response_content,
-    get_openai_usage, openai_response_to_message,
+use crate::providers::formats::openai::{
+    create_request, get_usage, is_context_length_error, response_to_message,
 };
 use mcp_core::tool::Tool;
 
@@ -77,28 +74,21 @@ impl Provider for OpenRouterProvider {
 
     #[tracing::instrument(
         skip(self, system, messages, tools),
-        fields(
-            model_config,
-            input,
-            output,
-            input_tokens,
-            output_tokens,
-            total_tokens,
-            cost
-        )
+        fields(model_config, input, output, input_tokens, output_tokens, total_tokens)
     )]
-    async fn complete_internal(
+    async fn complete(
         &self,
         system: &str,
         messages: &[Message],
         tools: &[Tool],
     ) -> Result<(Message, ProviderUsage)> {
         // Create the base payload
-        let payload = create_openai_request_payload_with_concat_response_content(
+        let payload = create_request(
             &self.model,
             system,
             messages,
             tools,
+            &super::utils::ImageFormat::OpenAi,
         )?;
 
         // Make request
@@ -106,123 +96,21 @@ impl Provider for OpenRouterProvider {
 
         // Raise specific error if context length is exceeded
         if let Some(error) = response.get("error") {
-            if let Some(err) = check_openai_context_length_error(error) {
+            if let Some(err) = is_context_length_error(error) {
                 return Err(err.into());
             }
             return Err(anyhow!("OpenRouter API error: {}", error));
         }
 
         // Parse response
-        let message = openai_response_to_message(response.clone())?;
+        let message = response_to_message(response.clone())?;
         let usage = self.get_usage(&response)?;
         let model = get_model(&response);
-        let cost = cost(&usage, &model_pricing_for(&model));
-        emit_debug_trace(self, &payload, &response, &usage, cost);
-        Ok((message, ProviderUsage::new(model, usage, cost)))
+        emit_debug_trace(self, &payload, &response, &usage);
+        Ok((message, ProviderUsage::new(model, usage)))
     }
 
     fn get_usage(&self, data: &Value) -> Result<Usage> {
-        get_openai_usage(data)
-    }
-}
-
-#[async_trait]
-impl Moderation for OpenRouterProvider {
-    async fn moderate_content(&self, _content: &str) -> Result<ModerationResult> {
-        Ok(ModerationResult::new(false, None, None))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::message::MessageContent;
-    use crate::providers::configs::ModelConfig;
-    use crate::providers::mock_server::{
-        create_mock_open_ai_response, create_mock_open_ai_response_with_tools, create_test_tool,
-        get_expected_function_call_arguments, setup_mock_server, TEST_INPUT_TOKENS,
-        TEST_OUTPUT_TOKENS, TEST_TOOL_FUNCTION_NAME, TEST_TOTAL_TOKENS,
-    };
-    use rust_decimal_macros::dec;
-    use wiremock::MockServer;
-
-    async fn _setup_mock_response(response_body: Value) -> (MockServer, OpenRouterProvider) {
-        let mock_server = setup_mock_server("/api/v1/chat/completions", response_body).await;
-
-        let provider = OpenRouterProvider {
-            client: Client::builder().build().unwrap(),
-            host: mock_server.uri(),
-            api_key: "test_api_key".to_string(),
-            model: ModelConfig::new("gpt-3.5-turbo".to_string()).with_temperature(Some(0.7)),
-        };
-
-        (mock_server, provider)
-    }
-
-    #[tokio::test]
-    async fn test_complete_basic() -> Result<()> {
-        let model_name = "gpt-4";
-        // Mock response for normal completion
-        let response_body =
-            create_mock_open_ai_response(model_name, "Hello! How can I assist you today?");
-
-        let (_, provider) = _setup_mock_response(response_body).await;
-
-        // Prepare input messages
-        let messages = vec![Message::user().with_text("Hello?")];
-
-        // Call the complete method
-        let (message, usage) = provider
-            .complete("You are a helpful assistant.", &messages, &[])
-            .await?;
-
-        // Assert the response
-        if let MessageContent::Text(text) = &message.content[0] {
-            assert_eq!(text.text, "Hello! How can I assist you today?");
-        } else {
-            panic!("Expected Text content");
-        }
-        assert_eq!(usage.usage.input_tokens, Some(TEST_INPUT_TOKENS));
-        assert_eq!(usage.usage.output_tokens, Some(TEST_OUTPUT_TOKENS));
-        assert_eq!(usage.usage.total_tokens, Some(TEST_TOTAL_TOKENS));
-        assert_eq!(usage.model, model_name);
-        assert_eq!(usage.cost, Some(dec!(0.00018)));
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_complete_tool_request() -> Result<()> {
-        // Mock response for tool calling
-        let response_body = create_mock_open_ai_response_with_tools("gpt-4");
-
-        let (_, provider) = _setup_mock_response(response_body).await;
-
-        // Input messages
-        let messages = vec![Message::user().with_text("What's the weather in San Francisco?")];
-
-        // Call the complete method
-        let (message, usage) = provider
-            .complete(
-                "You are a helpful assistant.",
-                &messages,
-                &[create_test_tool()],
-            )
-            .await?;
-
-        // Assert the response
-        if let MessageContent::ToolRequest(tool_request) = &message.content[0] {
-            let tool_call = tool_request.tool_call.as_ref().unwrap();
-            assert_eq!(tool_call.name, TEST_TOOL_FUNCTION_NAME);
-            assert_eq!(tool_call.arguments, get_expected_function_call_arguments());
-        } else {
-            panic!("Expected ToolCall content");
-        }
-
-        assert_eq!(usage.usage.input_tokens, Some(TEST_INPUT_TOKENS));
-        assert_eq!(usage.usage.output_tokens, Some(TEST_OUTPUT_TOKENS));
-        assert_eq!(usage.usage.total_tokens, Some(TEST_TOTAL_TOKENS));
-
-        Ok(())
+        get_usage(data)
     }
 }
