+++ crates/goose/tests/providers.rs
@@ -2,8 +2,75 @@ use anyhow::Result;
 use dotenv::dotenv;
 use goose::message::{Message, MessageContent};
 use goose::providers::base::Provider;
-use goose::providers::{databricks, ollama, openai};
+use goose::providers::{anthropic, databricks, google, groq, ollama, openai, openrouter};
+use mcp_core::content::Content;
 use mcp_core::tool::Tool;
+use std::collections::HashMap;
+use std::sync::Arc;
+use std::sync::Mutex;
+
+#[derive(Debug, Clone, Copy)]
+enum TestStatus {
+    Passed,
+    Skipped,
+    Failed,
+}
+
+impl std::fmt::Display for TestStatus {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            TestStatus::Passed => write!(f, "✅"),
+            TestStatus::Skipped => write!(f, "⏭️"),
+            TestStatus::Failed => write!(f, "❌"),
+        }
+    }
+}
+
+struct TestReport {
+    results: Mutex<HashMap<String, TestStatus>>,
+}
+
+impl TestReport {
+    fn new() -> Arc<Self> {
+        Arc::new(Self {
+            results: Mutex::new(HashMap::new()),
+        })
+    }
+
+    fn record_status(&self, provider: &str, status: TestStatus) {
+        let mut results = self.results.lock().unwrap();
+        results.insert(provider.to_string(), status);
+    }
+
+    fn record_pass(&self, provider: &str) {
+        self.record_status(provider, TestStatus::Passed);
+    }
+
+    fn record_skip(&self, provider: &str) {
+        self.record_status(provider, TestStatus::Skipped);
+    }
+
+    fn record_fail(&self, provider: &str) {
+        self.record_status(provider, TestStatus::Failed);
+    }
+
+    fn print_summary(&self) {
+        println!("\n============== Providers ==============");
+        let results = self.results.lock().unwrap();
+        let mut providers: Vec<_> = results.iter().collect();
+        providers.sort_by(|a, b| a.0.cmp(b.0));
+
+        for (provider, status) in providers {
+            println!("{} {}", status, provider);
+        }
+        println!("=======================================\n");
+    }
+}
+
+lazy_static::lazy_static! {
+    static ref TEST_REPORT: Arc<TestReport> = TestReport::new();
+    static ref ENV_LOCK: Mutex<()> = Mutex::new(());
+}
 
 /// Generic test harness for any Provider implementation
 struct ProviderTester {
@@ -59,22 +126,66 @@ impl ProviderTester {
 
         let message = Message::user().with_text("What's the weather like in San Francisco?");
 
-        let (response, _) = self
+        let (response1, _) = self
             .provider
             .complete(
                 "You are a helpful weather assistant.",
-                &[message],
-                &[weather_tool],
+                &[message.clone()],
+                &[weather_tool.clone()],
             )
             .await?;
 
         // Verify we got a tool request
         assert!(
-            response
+            response1
                 .content
                 .iter()
                 .any(|content| matches!(content, MessageContent::ToolRequest(_))),
             "Expected tool request in response"
+        );
+
+        dbg!(&response1);
+
+        let id = &response1
+            .content
+            .iter()
+            .filter_map(|message| message.as_tool_request())
+            .last()
+            .expect("got tool request")
+            .id;
+
+        let weather = Message::user().with_tool_response(
+            id,
+            Ok(vec![Content::text(
+                "
+                  50°F°C
+                  Precipitation: 0%
+                  Humidity: 84%
+                  Wind: 2 mph
+                  Weather
+                  Saturday 9:00 PM
+                  Clear",
+            )]),
+        );
+
+        // Verify we construct a valid payload including the request/response pair for the next inference
+        let (response2, _) = self
+            .provider
+            .complete(
+                "You are a helpful weather assistant.",
+                &[message, response1, weather],
+                &[weather_tool],
+            )
+            .await?;
+
+        dbg!(&response2);
+
+        assert!(
+            response2
+                .content
+                .iter()
+                .any(|content| matches!(content, MessageContent::Text(_))),
+            "Expected text for final response"
         );
 
         Ok(())
@@ -94,67 +205,178 @@ fn load_env() {
     }
 }
 
-#[tokio::test]
-async fn test_openai_provider() -> Result<()> {
+/// Helper function to run a provider test with proper error handling and reporting
+async fn test_provider<F, T>(
+    name: &str,
+    required_vars: &[&str],
+    env_modifications: Option<HashMap<&str, Option<String>>>,
+    provider_fn: F,
+) -> Result<()>
+where
+    F: FnOnce() -> Result<T>,
+    T: Provider + Send + Sync + 'static,
+{
+    // Take exclusive access to environment modifications
+    let lock = ENV_LOCK.lock().unwrap();
+
     load_env();
 
-    // Skip if credentials aren't available
-    if std::env::var("OPENAI_API_KEY").is_err() {
-        println!("Skipping OpenAI tests - credentials not configured");
+    // Save current environment state for required vars and modified vars
+    let mut original_env = HashMap::new();
+    for &var in required_vars {
+        if let Ok(val) = std::env::var(var) {
+            original_env.insert(var, val);
+        }
+    }
+    if let Some(mods) = &env_modifications {
+        for &var in mods.keys() {
+            if let Ok(val) = std::env::var(var) {
+                original_env.insert(var, val);
+            }
+        }
+    }
+
+    // Apply any environment modifications
+    if let Some(mods) = &env_modifications {
+        for (&var, value) in mods.iter() {
+            match value {
+                Some(val) => std::env::set_var(var, val),
+                None => std::env::remove_var(var),
+            }
+        }
+    }
+
+    // Setup the provider
+    let missing_vars = required_vars.iter().any(|var| std::env::var(var).is_err());
+    let provider = provider_fn();
+
+    // Restore original environment
+    for (&var, value) in original_env.iter() {
+        std::env::set_var(var, value);
+    }
+    if let Some(mods) = env_modifications {
+        for &var in mods.keys() {
+            if !original_env.contains_key(var) {
+                std::env::remove_var(var);
+            }
+        }
+    }
+
+    std::mem::drop(lock);
+
+    if missing_vars {
+        println!("Skipping {} tests - credentials not configured", name);
+        TEST_REPORT.record_skip(name);
         return Ok(());
     }
 
-    let provider = openai::OpenAiProvider::from_env()?;
-    let tester = ProviderTester::new(provider);
-    tester.run_test_suite().await?;
+    if provider.is_err() {
+        println!("Could not setup {} from env", name);
+        TEST_REPORT.record_fail(name);
+        return Err(provider.err().expect("is error"));
+    }
 
-    Ok(())
+    let tester = ProviderTester::new(provider.expect("already checked"));
+    match tester.run_test_suite().await {
+        Ok(_) => {
+            TEST_REPORT.record_pass(name);
+            Ok(())
+        }
+        Err(e) => {
+            println!("{} test failed: {}", name, e);
+            TEST_REPORT.record_fail(name);
+            Err(e)
+        }
+    }
+}
+
+#[tokio::test]
+async fn test_openai_provider() -> Result<()> {
+    test_provider(
+        "OpenAI",
+        &["OPENAI_API_KEY"],
+        None,
+        openai::OpenAiProvider::from_env,
+    )
+    .await
 }
 
 #[tokio::test]
 async fn test_databricks_provider() -> Result<()> {
-    load_env();
-
-    // Skip if credentials aren't available
-    if std::env::var("DATABRICKS_HOST").is_err() || std::env::var("DATABRICKS_TOKEN").is_err() {
-        println!("Skipping Databricks tests - credentials not configured");
-        return Ok(());
-    }
-
-    let provider = databricks::DatabricksProvider::from_env()?;
-    let tester = ProviderTester::new(provider);
-    tester.run_test_suite().await?;
-
-    Ok(())
+    test_provider(
+        "Databricks",
+        &["DATABRICKS_HOST", "DATABRICKS_TOKEN"],
+        None,
+        databricks::DatabricksProvider::from_env,
+    )
+    .await
 }
 
 #[tokio::test]
 async fn test_databricks_provider_oauth() -> Result<()> {
-    load_env();
+    let mut env_mods = HashMap::new();
+    env_mods.insert("DATABRICKS_TOKEN", None);
 
-    // Skip if credentials aren't available
-    if std::env::var("DATABRICKS_HOST").is_err() {
-        println!("Skipping Databricks OAuth tests - credentials not configured");
-        return Ok(());
-    }
-
-    // Clear any token to force OAuth
-    std::env::remove_var("DATABRICKS_TOKEN");
-    let provider = databricks::DatabricksProvider::from_env()?;
-    let tester = ProviderTester::new(provider);
-    tester.run_test_suite().await?;
-
-    Ok(())
+    test_provider(
+        "Databricks OAuth",
+        &["DATABRICKS_HOST"],
+        Some(env_mods),
+        databricks::DatabricksProvider::from_env,
+    )
+    .await
 }
 
-// Integration tests that run against a real Ollama server
 #[tokio::test]
 async fn test_ollama_provider() -> Result<()> {
-    load_env();
+    test_provider("Ollama", &[], None, ollama::OllamaProvider::from_env).await
+}
 
-    let provider = ollama::OllamaProvider::from_env()?;
-    let tester = ProviderTester::new(provider);
-    tester.run_test_suite().await?;
+#[tokio::test]
+async fn test_groq_provider() -> Result<()> {
+    test_provider(
+        "Groq",
+        &["GROQ_API_KEY"],
+        None,
+        groq::GroqProvider::from_env,
+    )
+    .await
+}
 
-    Ok(())
+#[tokio::test]
+async fn test_anthropic_provider() -> Result<()> {
+    test_provider(
+        "Anthropic",
+        &["ANTHROPIC_API_KEY"],
+        None,
+        anthropic::AnthropicProvider::from_env,
+    )
+    .await
+}
+
+#[tokio::test]
+async fn test_openrouter_provider() -> Result<()> {
+    test_provider(
+        "OpenRouter",
+        &["OPENROUTER_API_KEY"],
+        None,
+        openrouter::OpenRouterProvider::from_env,
+    )
+    .await
+}
+
+#[tokio::test]
+async fn test_google_provider() -> Result<()> {
+    test_provider(
+        "Google",
+        &["GOOGLE_API_KEY"],
+        None,
+        google::GoogleProvider::from_env,
+    )
+    .await
+}
+
+// Print the final test report
+#[ctor::dtor]
+fn print_test_report() {
+    TEST_REPORT.print_summary();
 }
